extracts authentication into a seperate concern
